Garage/web-ui/requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
Jinja2==3.0.1
itsdangerous==2.0.1
MarkupSafe==2.0.1
psutil==5.9.0
numpy==1.21.0
matplotlib==3.4.2
gunicorn==20.1.0
pillow==8.2.0
flask-caching==1.10.1